requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.0.1
python-dotenv==1.0.0
